pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "research-pipeline"
version = "0.1.0"
description = "Local PDF research pipeline with evidence-linked reports."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai>=1.30.0",
  "pydantic>=2.6.0",
  "pyyaml>=6.0.1",
  "typer>=0.12.0",
]

[project.scripts]
research = "research_pipeline.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
